requirements.txt
yfinance
requests
pandas
groq
python-dotenv
pytz
